feat(reflex): Complete Phase 2 infrastructure - enhanced server with Redis
Implemented comprehensive server infrastructure with Redis connection pooling:

**Enhanced Main Server** (main.rs - 253 lines):
- Configuration loading from environment variables
- Configurable logging (JSON/text, multiple levels)
- Graceful shutdown handling (SIGTERM/SIGINT)
- Health endpoint with version and uptime
- Readiness endpoint with Redis connectivity check
- Metrics endpoint placeholder
- TraceLayer and CORS middleware
- AppState with Redis client integration

**Redis Client Module** (redis_client.rs - 204 lines):
- deadpool-redis connection pooling (configurable pool size)
- Exponential backoff retry logic (3 retries, 100ms-5s backoff)
- Health check with PING command
- Pool status monitoring
- Connection timeout configuration
- Comprehensive unit tests (8 tests, 2 ignored integration tests)

**Dependencies Added**:
- chrono 0.4 (timestamps)
- deadpool 0.12 (generic pooling)
- deadpool-redis 0.18 (Redis pooling)

**Configuration**:
- All settings configurable via REFLEX_* environment variables
- Sensible defaults for all parameters
- Redis URL, pool size, timeouts, cache TTL

**Test Results**:
- 10 total tests: 8 passed, 2 ignored (integration)
- All unit tests passing
- Integration tests require running Redis instance

+++ services/reflex-layer/src/config.rs
@@ -4,7 +4,6 @@
 //! Supports environment-based overrides for development, staging, and production.
 
 use serde::Deserialize;
-use std::env;
 use std::time::Duration;
 
 /// Main configuration structure for the Reflex Layer
@@ -153,7 +152,7 @@ impl Config {
 
             // Performance defaults
             .set_default("performance.max_concurrent_requests", 1000)?
-            .set_default("performance.worker_threads", num_cpus::get())?
+            .set_default("performance.worker_threads", num_cpus::get() as i64)?
 
             // Logging defaults
             .set_default("logging.level", "info")?
